Initial commit
Naive approach. Very inefficient. Essentially just iterates on a really
large range checking every multiple of 5 to see if it is evenly
divisible by every number 1-20. One easy way to speed this up would be
to check on the rules of divisibility testing and see which of these
conditions can be removed (example: something is divisible by 10 if it
is divisible by 2 and 5. So if I have 2 and 5 in there I don’t need 10.
Or maybe if I have 10 and I don’t need 2 or 5).